backend/requirements.txt
fastapi==0.116.2
starlette>=0.27.0
uvicorn==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
redis>=5.0.0
stripe==12.5.1
pydantic-settings==2.10.1
python-dotenv==1.1.1
email-validator>=2.0.0
psycopg2-binary
websockets
google-generativeai
groq>=0.12.0
alembic
Pillow>=10.0.0
psutil>=7.1.3
python-multipart>=0.0.5
aiosmtplib>=2.1.0
email-validator>=2.0.0
